requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.10
packaging==24.1
pillow==11.0.0
pystray==0.19.5
python-xlib==0.33
requests==2.32.3
six==1.16.0
typing==3.7.4.3
urllib3==2.2.3
windows-tools.installed-software==0.5.4
windows-tools.misc==1.0.0
windows-tools.registry==1.1.0
winregistry==1.1.1
